requirements.txt
google-adk
google-adk[a2a]
python_dotenv